Remove fields element from form class
With a little wiggling, we can now access fields by their member name, so
form.email is how we access the email field. The only other purpose (real
purpose) of the fields element in the Form class was to specify the order we
should draw the fields in. Custom templates means we can arrange our fields in
the template instead, at the cost of a little file bloat that's probably going
to prove a wash once we start skinning the site.

Signed-off-by: Casey Dahlin <casey.dahlin@gmail.com>